Added totalPrice and boughtProductCounter updater from database

+++ app/src/main/java/com/tommyapps/shop/Product.java
@@ -33,7 +33,7 @@ public class Product {
         return price;
     }
 
-    public void setPrice(float price) {
+    public void setPrice(double price) {
         this.price = price;
     }
